Move parseGitHubAuthorizationHeader and IGitHubAuthorizationHeader to GitHubClient; add GitHubClient tests
